+++ internal/ui/viewmodel/editor_viewmodel.go
@@ -185,7 +185,7 @@ func (v *editorViewModelImpl) handleFileSave(e event.Event) {
 			}))
 			return
 		}
-		pl.File.SetSizeBytes(len(pl.Content))
+		pl.File.SetSizeBytes(uint64(len(pl.Content)))
 		v.bus.Publish(e.NewFollowup(fileeditor.SaveSucceeded(pl)))
 	}()
